codegen-inkwell: add IRRepresentable trait
and reorganise code hierarchy
